In this version, the user can upload multiple images at the same time.

+++ app/controllers/exams_controller.rb
@@ -3,20 +3,46 @@ class ExamsController < ApplicationController
     @exams = Exam.all
   end
 
-  def new
-    @exam = Exam.new
+  def show
+    @exam = Exam.find(params[:id])
+    @pages = @exam.pages.all
   end
 
+  def new
+    @exam = Exam.new
+    @page = @exam.pages.build
+  end
+
+  # This method stores an exam and associates it to a pdf file in case the file
+  # had been successfully created.
   def create
     @exam = Exam.new(exam_params)
-    exam_pdf_path = create_pdf(@exam)
-    @exam.pdf = File.open(exam_pdf_path)
-    if @exam.save
-      File.delete(exam_pdf_path)
-      redirect_to exams_path, notice: "El examen \"#{@exam.name}\" se ha 
-      subido."
+    if params[:pages]
+      if @exam.save
+        params[:pages]['image'].each do |p|
+          @exam.pages.create!(:image => p, :exam_id => @exam.id)
+        end
+        exam_pdf_path = create_pdf(@exam)
+        @exam.pdf = File.open(exam_pdf_path)
+        if @exam.save
+          File.delete(exam_pdf_path)
+          redirect_to exams_path, notice: "El examen \"#{@exam.name}\" se ha 
+          subido."
+        else
+          @exam.destroy
+          render json: {
+            success: false,
+            errors: "No fue posible guardar el archivo."
+          }
+        end
+      else
+        render "new"
+      end
     else
-      render "new"
+      render json: {
+        success: false,
+        errors: "Se debe adjuntar al menos una imagen."
+      }
     end
   end
 
@@ -28,23 +54,32 @@ class ExamsController < ApplicationController
 
   private
     def exam_params
-      params.require(:exam).permit(:name, :attachment)
+      params.require(:exam).permit(:name, pages_attributes: [:number, :image])
     end
 
+    # This method writes the images array to a pdf file and returns its 
+    # location path.
     def create_pdf(exam)
-      exam_img_list = Magick::ImageList.new(exam.attachment.path())
       exam_pdf_path = Rails.root.join('public', 'uploads', 'tmp', 'pdfs', 
                                       exam.name + '.pdf')
       create_non_existing_dirs(exam_pdf_path)
-      exam_img_list[0].write(exam_pdf_path)
+      images_array = Array.new
+      exam.pages.each do |p|
+        images_array.push p.image.path()
+      end
+      image_list = Magick::ImageList.new(*images_array)
+      image_list.write(exam_pdf_path)
+
       return exam_pdf_path
     end
 
+    # This method creates any non existing directory in the pdf path so that the 
+    #  file can be stored.
     def create_non_existing_dirs(path)
       require 'fileutils'
-      dirname = File.dirname(path)
-      unless File.directory?(dirname)
-        FileUtils.mkdir_p(dirname)
+      dir_name = File.dirname(path)
+      unless File.directory?(dir_name)
+        FileUtils.mkdir_p(dir_name)
       end
     end
 end
